Création des modules Eco-PTZ et PAR+ pour intégration iframe
